@@ -1,3 +1,3 @@
-@using PersonalPhotos
+﻿@using PersonalPhotos
 @using PersonalPhotos.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
